Initial work on breaking 0.10 version
By introducing states that must be dereferenced to access its value, the
need to alter-var-root the value once started disappears. Because of
this, there is more control over what value to return. This is
leveraged by having thread bound states. This way, multiple systems can
co-exist, which can - for example - be used for parallel testing.