Improve extra fields in defstates
Extra fields from substitutes are inserted when started, and are kept
inside the sessions data.

Ensure that extra fields dissappear from a defstate when those are
removed when re-evaluated.

Add an undocumented as-started function, that returns a merge of the
defstate fields and the state fields how it was started.