Order transitive dependencies
Currently, transitive dependencies are being started in the order they
were loaded, not in dependency order. This can cause issues in
situations where new states which are needed for substitutes are
loaded much later than the original state that is being overridden to
depend on those new states.

An example of this is in a test environment where you may have
configuration overrides that you wouldn't define in your production
overrides that are loaded from dev-only code. Without this change, the
new test that was written failed because the newly added dependency was
not loaded first.

I had to do a bit of a hack to remove a state defined by defstate-let me
know if there is a better way of doing this.