Fix redefining autostarting state (#33)
The autostart extension works by wrapping a State record by the
AutoStartState record, by updating the var created by the standard
defstate. Since that var is a defonce, the autostart extension wrapped
itself on redefining a state.

The fix checks whether the var already exists and is an
AutoStartState. If so, it first unwraps it.